pyproject.toml
[project]
name = "stock-portfolio-tracker-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "celery>=5.5.3",
    "dotenv>=0.9.9",
    "fastapi>=0.117.1",
    "httpx==0.27.2",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.9",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
